+++ src/app/admin-module/models/checkbox-list-service.interface.ts
@@ -1,0 +1,7 @@
+export interface CheckboxListService {
+  /**
+   * Function to return label for checkbox
+   * @param listItem List item
+   */
+  getLabel(listItem: unknown): string;
+}