+++ src/app/language-switch-module/models/language-object.type.ts
@@ -1,0 +1,9 @@
+import { Language } from './language.type';
+
+type LanguageValue = {
+  [key: string]: string;
+};
+/**
+ * Type of language objects used inside component to define easy localization texts.
+ */
+export type LanguageObject = Record<Language, LanguageValue>;