+++ src/app/search-module/components/search-bar/search-bar.component.ts
@@ -2,7 +2,7 @@ import { Component, OnDestroy, OnInit, ViewEncapsulation } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Subject, takeUntil } from 'rxjs';
 
-import { IngredientService } from '../../services/ingredient.service';
+import { IngredientService } from '../../../shared/services/ingredient-service/ingredient.service';
 import locale from './search-bar.locale.json';
 import { IngredientCategory } from '../../models/ingredient-category.interface';
 import { CategoryService } from '../../services/category.service';